move webpack postwriting microservices to an abstract uberclass on top of client technologies
