+++ generators/vue/generator.mjs
@@ -65,11 +65,11 @@ export default class VueGenerator extends AbstractJHipsterClientGenerator {
   }
 
   get loading() {
-    return this.asLoadingTaskGroup({
+    return {
       loadPackageJson() {
         _.merge(this.dependabotPackageJson, this.fs.readJSON(this.fetchFromInstalledJHipster('vue', 'templates', 'package.json')));
       },
-    });
+    };
   }
 
   get [BaseApplicationGenerator.LOADING]() {
@@ -77,11 +77,11 @@ export default class VueGenerator extends AbstractJHipsterClientGenerator {
   }
 
   get preparing() {
-    return this.asPreparingTaskGroup({
+    return {
       prepareForTemplates({ application }) {
         application.webappEnumerationsDir = `${application.clientSrcDir}app/shared/model/enumerations/`;
       },
-    });
+    };
   }
 
   get [BaseApplicationGenerator.PREPARING]() {
@@ -96,7 +96,7 @@ export default class VueGenerator extends AbstractJHipsterClientGenerator {
   }
 
   get [BaseApplicationGenerator.WRITING]() {
-    return this.delegateTasksToBlueprint(() => this.writing);
+    return this.asWritingTaskGroup(this.delegateTasksToBlueprint(() => this.writing));
   }
 
   get writingEntities() {
@@ -107,7 +107,24 @@ export default class VueGenerator extends AbstractJHipsterClientGenerator {
   }
 
   get [BaseApplicationGenerator.WRITING_ENTITIES]() {
-    return this.delegateTasksToBlueprint(() => this.writingEntities);
+    return this.asWritingEntitiesTaskGroup(this.delegateTasksToBlueprint(() => this.writingEntities));
+  }
+
+  /**
+   * Postwriting step for client technology
+   * @return {{microfrontend({application: *}): void}}
+   */
+  get postWriting() {
+    return {
+      microfrontend({ application }) {
+        if (!application.microfrontend) return;
+        this.addWebpackConfig("require('./webpack.microfrontend')({ serve: options.env.WEBPACK_SERVE })", application.clientFramework);
+      },
+    };
+  }
+
+  get [BaseApplicationGenerator.POST_WRITING]() {
+    return this.asPostWritingTaskGroup(this.delegateTasksToBlueprint(() => this.postWriting));
   }
 
   get postWritingEntities() {
@@ -117,7 +134,7 @@ export default class VueGenerator extends AbstractJHipsterClientGenerator {
   }
 
   get [BaseApplicationGenerator.POST_WRITING_ENTITIES]() {
-    return this.delegateTasksToBlueprint(() => this.postWritingEntities);
+    return this.asPostWritingEntitiesTaskGroup(this.delegateTasksToBlueprint(() => this.postWritingEntities));
   }
 
   /**
