+++ test/needle-api/needle-client-webpack-react.spec.mts
@@ -4,13 +4,15 @@ import ClientGenerator from '../../generators/client/index.mjs';
 import { CLIENT_WEBPACK_DIR } from '../../generators/generator-constants.mjs';
 import { clientFrameworkTypes } from '../../jdl/jhipster/index.mjs';
 import { getGenerator } from '../support/index.mjs';
+import AngularGenerator from '../../generators/angular/index.mjs';
+import ReactGenerator from '../../generators/react/index.mjs';
 
 const { ANGULAR, VUE, REACT } = clientFrameworkTypes;
 const assetFrom = 'source';
 const assetTo = 'target';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-const mockBlueprintSubGen: any = class extends ClientGenerator {
+const mockBlueprintSubGen: any = class extends ReactGenerator {
   constructor(args, opts, features) {
     super(args, opts, features);
 
@@ -34,10 +36,10 @@ const mockBlueprintSubGen: any = class extends ClientGenerator {
   }
 };
 
-describe('needle API Webpack: JHipster client generator with blueprint', () => {
+describe('needle API Webpack - React: JHipster client generator with blueprint', () => {
   function generateAppWithClientFramework(clientFramework) {
     return helpers
-      .create(getGenerator('client'))
+      .create(getGenerator('react'))
       .withOptions({
         build: 'maven',
         auth: 'jwt',
@@ -52,40 +54,13 @@ describe('needle API Webpack: JHipster client generator with blueprint', () => {
         nativeLanguage: 'en',
         languages: ['en', 'fr'],
       })
-      .withGenerators([[mockBlueprintSubGen, 'jhipster-myblueprint:client']])
+      .withGenerators([[mockBlueprintSubGen, 'jhipster-myblueprint:react']])
       .run();
   }
-
-  describe('Angular clientFramework', () => {
-    before(() => {
-      return generateAppWithClientFramework(ANGULAR);
-    });
-
-    it('Assert external asset is added to webpack.custom.js', async () => {
-      assert.fileContent(`${CLIENT_WEBPACK_DIR}webpack.custom.js`, `{ from: '${assetFrom}', to: '${assetTo}' },`);
-    });
-
-    it('should add webpack config to webpack.custom.js', async () => {
-      assert.fileContent(`${CLIENT_WEBPACK_DIR}webpack.custom.js`, '{ devServer: {} }');
-    });
-  });
 
   describe('React clientFramework', () => {
     before(() => {
       return generateAppWithClientFramework(REACT);
-    });
-
-    it('Assert external asset is added to webpack.common.js', async () => {
-      assert.fileContent(`${CLIENT_WEBPACK_DIR}webpack.common.js`, `{ from: '${assetFrom}', to: '${assetTo}' },`);
-    });
-    it('should add webpack config to webpack.common.js', async () => {
-      assert.fileContent(`${CLIENT_WEBPACK_DIR}webpack.common.js`, '{ devServer: {} }');
-    });
-  });
-
-  describe('Vue clientFramework', () => {
-    before(() => {
-      return generateAppWithClientFramework(VUE);
     });
 
     it('Assert external asset is added to webpack.common.js', async () => {
