+++ test/needle-api/needle-client-webpack-vue.spec.mts
@@ -1,0 +1,73 @@
+import assert from 'yeoman-assert';
+import helpers from 'yeoman-test';
+import ClientGenerator from '../../generators/client/index.mjs';
+import { CLIENT_WEBPACK_DIR } from '../../generators/generator-constants.mjs';
+import { clientFrameworkTypes } from '../../jdl/jhipster/index.mjs';
+import { getGenerator } from '../support/index.mjs';
+import AngularGenerator from '../../generators/angular/index.mjs';
+import VueGenerator from '../../generators/vue/index.mjs';
+
+const { ANGULAR, VUE, REACT } = clientFrameworkTypes;
+const assetFrom = 'source';
+const assetTo = 'target';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const mockBlueprintSubGen: any = class extends VueGenerator {
+  constructor(args, opts, features) {
+    super(args, opts, features);
+
+    const jhContext = (this.jhipsterContext = this.options.jhipsterContext);
+
+    if (!jhContext) {
+      throw new Error('This is a JHipster blueprint and should be used only like jhipster --blueprints myblueprint');
+    }
+
+    this.sbsBlueprint = true;
+  }
+
+  get [ClientGenerator.POST_WRITING]() {
+    const customPhaseSteps = {
+      webpackPhase() {
+        this.copyExternalAssetsInWebpack(assetFrom, assetTo);
+        this.addWebpackConfig('{devServer:{}}');
+      },
+    };
+    return { ...customPhaseSteps };
+  }
+};
+
+describe('needle API Webpack - Vue: JHipster client generator with blueprint', () => {
+  function generateAppWithClientFramework(clientFramework) {
+    return helpers
+      .create(getGenerator('vue'))
+      .withOptions({
+        build: 'maven',
+        auth: 'jwt',
+        db: 'mysql',
+        skipInstall: true,
+        blueprint: 'myblueprint',
+        skipChecks: true,
+        skipServer: true,
+        baseName: 'jhipster',
+        clientFramework,
+        enableTranslation: true,
+        nativeLanguage: 'en',
+        languages: ['en', 'fr'],
+      })
+      .withGenerators([[mockBlueprintSubGen, 'jhipster-myblueprint:vue']])
+      .run();
+  }
+
+  describe('Vue clientFramework', () => {
+    before(() => {
+      return generateAppWithClientFramework(VUE);
+    });
+
+    it('Assert external asset is added to webpack.common.js', async () => {
+      assert.fileContent(`${CLIENT_WEBPACK_DIR}webpack.common.js`, `{ from: '${assetFrom}', to: '${assetTo}' },`);
+    });
+    it('should add webpack config to webpack.common.js', async () => {
+      assert.fileContent(`${CLIENT_WEBPACK_DIR}webpack.common.js`, '{ devServer: {} }');
+    });
+  });
+});
